implement TeamListings component
